+++ packages/fetch/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @sthirajs/fetch
 
+## 2.0.0
+
+### Patch Changes
+
+- Updated dependencies
+  - @sthirajs/core@0.3.0
+
 ## 1.1.0
 
 ### Minor Changes